Some changes on NEMS GFS are ready to commit to repository. These changes are:
1) The 6 hour outputs for NEMS GFS will be the filtered variables when
the digital filter is turned on. Current NEMS GFS doesn't rewrite the 6
hour outputs when digital filter is turned on because ESMF alarm only
rings once. Changes are made in src/main/atmos/module_GFS_INTEGRATE.F90
on the ESMF alarm to write out  filtered  variables.

2) A header array varr8val in dynamics restart file will be saved in
src/gfs_dyn/twriteg_rst.f.

3) The compiler option for gfs_libutil is set to -qrealsize=8, current
compiler for gfs_libutil has no this option, which means it is using
default real4 for the real variables in gfs_libutil, and this is not
consistent with gfs_dyn and gfs_phys. Specifically, the real variables
in src/gfs_libutil/module_digital_filter_gfs.F90 is hardcoded as real8
since the Fortran array obtained from ESMF state is ESMF_KIND_R8.

Due to these changes, the regression test output for NEMS GFS with
digital filter turned on will be changed.  Regression test passed with
the updated outputs.

No impact on NEMS NMMB code.

M      ush/compile_options/gen_compile_options_Linux
M      ush/compile_options/gen_compile_options_AIX
M      src/gfs_libutil/module_digital_filter_gfs.F90
M      src/main/atmos/module_GFS_INTEGRATE.F90
M      src/gfs_dyn/twriteg_rst.f
M      src/gfs_dyn/twrites_rst.f
M      src/gfs_dyn/gfs_dyn_dfi_mod.f
M      job/regression_tests/RegressionTests.log


git-svn-id: https://svnemc.ncep.noaa.gov/projects/nems/trunk@6849 281fb692-3a18-0410-826d-9e847983f365

+++ src/gfs_libutil/module_digital_filter_gfs.F90
@@ -13,8 +13,8 @@
 ! ---------
 ! dynamics
 ! ---------
-      real, allocatable, save :: dyn_array_save(:,:,:,:)
-      real             , save :: totalsum
+      real(8), allocatable, save :: dyn_array_save(:,:,:,:)
+      real(8)             , save :: totalsum
       character(20), allocatable, save :: dyn_name(:)
       integer,       allocatable, save :: dyn_dim(:,:)
       integer,       allocatable, save :: dyn_items_ord(:)
@@ -131,7 +131,7 @@
       TYPE(ESMF_Field)		    :: Field
       real(ESMF_KIND_R8), dimension(:,:), pointer :: tmp_ptr2d
       real(ESMF_KIND_R8), dimension(:,:,:), pointer :: tmp_ptr
-      real                          :: sx, wx, digfil
+      real(8)                          :: sx, wx, digfil
       integer                       :: n, i, rc,item,dim1,dim2,dim3
 
         kstep = kstep + 1
@@ -189,7 +189,7 @@
       real(ESMF_KIND_R8), dimension(:,:,:), pointer :: tmp_ptr
       real(ESMF_KIND_R8), dimension(:,:), pointer :: tmp_ptr2D
       real(ESMF_KIND_R8), dimension(:,:,:), pointer :: tmp_ptr1
-      real                            :: totalsumi
+      real(8)                         :: totalsumi
       integer                         :: n, i, rc,item
       integer                         :: dim1,dim2,dim3
 !
